Low Latency BIOS configuration and validation tool
The correct configuration of the hardware where the
workloads of the different network functions are to
be executed is crucial to achieve low latency systems.
Most of this configuration is done through the system
BIOS attributes. Once we have understood the function
and impact of these attributes in our particular case
(i.e., what would be the most appropriate values of
these attributes to minimize the latency of our system),
the next question is to transfer these values to each
particular vendor.

The aim of this tool is helping to solve these two
problems as much as possible.